refactor: move core ics files to lib move graphics libraries into ics as well
