+++ sim/lib/core/include/core/graphicsContext.h
@@ -1,6 +1,8 @@
 #ifndef BENTOBOX_GRAPHICSCONTEXT_H
 #define BENTOBOX_GRAPHICSCONTEXT_H
 
+#include <string>
+#include <unordered_map>
 #include "windowContext.h"
 
 class GraphicsContext {
@@ -11,6 +13,9 @@ public:
     GraphicsContext& operator=(const GraphicsContext& other);
 
     WindowContext& windowContext;
+
+    std::unordered_map<std::string, unsigned int> textureCache;
+    unsigned int loadTexture2D(std::string const& filepath);
 };
 
 #endif //BENTOBOX_GRAPHICSCONTEXT_H
