+++ sim/lib/core/include/core/ics/util/typeMap.h
