+++ sim/lib/core/lib/glad/include/KHR/khrplatform.h
