+++ sim/lib/core/lib/stb_image/stb_image.h
