+++ sim/lib/core/src/windowContext.cpp
@@ -1,4 +1,4 @@
-#include "windowContext.h"
+#include <core/windowContext.h>
 
 namespace {
     void framebufferSizeCallback([[maybe_unused]] GLFWwindow* window, int width, int height) {
