--- sim/src/core/ics/CMakeLists.txt
@@ -1,30 +1,0 @@
-SET(LIB_NAME ics)
-SET(SOURCES component.h componentManager.h componentSet.h componentStore.h compVec.h index.h)
-SET(TESTS component.test.cpp componentManager.test.cpp compVec.test.cpp index.test.cpp)
-
-add_library(
-    ${LIB_NAME}
-    ${SOURCES}
-)
-
-add_library(
-    "${LIB_NAME}-test"
-    ${SOURCES}
-    ${TESTS}
-)
-
-# Cmake could not determine the language used from the files
-# https://stackoverflow.com/questions/11801186/cmake-unable-to-determine-linker-language-with-c
-set_target_properties(${LIB_NAME} PROPERTIES LINKER_LANGUAGE CXX)
-set_target_properties("${LIB_NAME}-test" PROPERTIES LINKER_LANGUAGE CXX)
-
-target_include_directories(${LIB_NAME} PRIVATE ".")
-target_include_directories("${LIB_NAME}-test" PUBLIC ".")
-
-# Util library
-add_subdirectory("${CMAKE_CURRENT_SOURCE_DIR}/util")
-target_link_libraries(${LIB_NAME} PUBLIC util)
-target_link_libraries("${LIB_NAME}-test" PUBLIC -Wl,--whole-archive,--no-as-needed util-test -Wl,--no-whole-archive,--no-as-needed)
-
-# Google test
-target_link_libraries("${LIB_NAME}-test" PRIVATE gtest)
